requirements.txt
streamlit
torch
torchvision
numpy
pillow
opencv-python
scikit-image